Separate create cbz to new element
